requirements.txt
pysubnettree
pyyaml
redis
stix2
tld
validators
